javítva, double az erdmény

+++ 3_feladat/src/Main.java
@@ -8,7 +8,7 @@ public class Main {
 		int d = Integer.parseInt(args[3]);
 		int e = Integer.parseInt(args[4]);
 
-		int avarage = (a + b + c + d + e) / 5;
+		double avarage = (double) (a + b + c + d + e) / 5;
 		System.out.println(
 				"a: " + args[0] + ", b: " + args[1] + ", c: " + args[2] + ", d: " + args[3] + ", e: " + args[4]);
 		System.out.println("A számok számtani közepe: " + avarage);
